Make the miniplayer available on all screens
